Added multiboot2 module support
This is not very complex of an addition, but it has
its quirks. First, there can be multiple module tags
in the multiboot info structure. Because of this a
ModuleIter would include a modified TagIter or TagIter
would need a get_next_tag(x) method. Secondly the tag
contiains a string at the end that I use an unchecked
conversion to access. I think this is fine though because
this method is already used in boot_loader_name.rs.