multiboot2: Fix EFI Memory Map 'last_area' calculation
Correct EFI 'last_area' calculation so it correctly describes the greatest possible starting
address of the last record.

(Previously, iterating the EFI Memory map resulted in a superfluous entry as it ran over the next tag)

This revision also cleans up the E820/EFI parsing to use 0-length arrays for the records
and to faciliate a more natural 'sizeof' operation on the Multiboot2 tag.