multiboot2: boxed_dst_tag: don't be too strict about allocation size
I don't exactly understand why, but running this in the integration test
revealed a different behaviour compared to the unit test. In the unit test
environment, it always seemed like the next multiple of 8 is used for each
allocation as size. In the integration test, I did not encounter this. I
don't know what's the best solution here. Probably, just don't be too
strict.