multiboot2: Follow C behavior for string parsing
The various strings in the Multiboot2 spec are specified as C-style
zero-terminated strings.

Previously, if the multiboot2 crate is handed a string with non-NUL characters
present between the terminating NUL (within the tag), they were returned as
well.

This revision instead terminates the strings on the first NUL character,
as would be typical in C handling. Although unterminated strings in Multiboot2
are technically an error, current unterminated string behavior is maintained.