multiboot2: massive internal refactoring + most tests pass Miri 🎉
The goal of this commit is to prepare 100% memory safety (using Miri
passing tests as reference). For that, we need a significant
under-the-hood changes.

`GenericTag` is the generalization of all tags as DST that owns all
memory of the tag. This tag can be created from bytes, thus, we can
ensure a lifetime and a valid memory range. This tag then can be
casted to specific implementations of `TagTrait`. We now never have to
increase or decrease the size of the referenced memory during a Tag
cast, as the GenericTag already holds all bytes that the more specific
type needs.

Assertions and the memory checks along the way ensure that nothing can
co wrong.

Further, the creation of various test data structures was modified to
fulfill the new guarantees, that are given in real-world scenarios and
are also what the compiler expects.